pyproject.toml
[project]
name = "oliveyoung-voice-assistant"
version = "0.1.0"
description = "올리브영 매장 정보 제공 음성 쇼핑 어시스턴트"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[openai,cartesia,silero,websocket]>=0.0.92",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.10.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

